multi: refactor and move HopPayload methods
This commit is a pure refactor and move commit. It moves the HopPayload
methods out from the `path.go` file and into a dedicated `payload.go`
file in order to prepare for a follow up commit that will add a blinded
path struct to the `path.go` file. The opportunity is also taken to
separate out the tlv vs legacy encoding of the hop payload a bit
(seprate constructors for example) in order to prepare for a future
where the legacy encoding will be removed.
